+++ packages/ui/badge/src/badgeIcon/BadgeIcon.tsx
@@ -1,4 +1,4 @@
-import { IconByName } from '@pikas-ui/icons';
+import { IconProps } from '@pikas-ui/icons';
 import type { CustomBadgeProps } from '../customBadge/index.js';
 import { CustomBadge } from '../customBadge/index.js';
 import { styled } from '@pikas-ui/styles';
@@ -28,18 +28,20 @@ const BadgeStyled = styled(CustomBadge, {
 });
 
 export type BadgeIconProps = CustomBadgeProps & {
-  iconName: string;
+  Icon: FC<IconProps>;
+  iconProps?: IconProps;
   size?: number | string;
   padding?: BadgePadding;
 };
 
 export const BadgeIcon: FC<BadgeIconProps> = ({
-  iconName,
+  Icon,
   size = 24,
   padding = 'md',
+  iconProps,
   ...props
 }) => (
   <BadgeStyled padding={padding} {...props}>
-    <IconByName name={iconName} size={size} />
+    <Icon size={size} {...iconProps} />
   </BadgeStyled>
 );